WIP removing Firebase Manager code.
